pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "mellea-skills-compiler"
description = "Agent specification certification pipeline — from skill specs to governed, certified Mellea pipelines"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.11, <3.14.4"
license = {text = "Apache-2.0"}
authors = [
    {name = "Elizabeth Daly", email = "elizabeth.daly@ie.ibm.com"},
    {name = "Dhaval Salwala", email = "dhaval.vinodbhai.salwala@ibm.com"},
    {name = "Inge Vejsbjerg", email = "ingevejs@ie.ibm.com"},
    {name = "Seshu Tirupathi", email = "seshutir@ie.ibm.com"},
    {name = "Rebecka Nordenlow", email = "Rebecka.Nordenlow@ibm.com"},
    {name = "Jessica He", email = "jessicahe@ibm.com"},
    {name = "Kush Varshney", email = "krvarshn@us.ibm.com"},
    {name = "Jordan McAfoose", email = "MCA@zurich.ibm.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
    "mellea[hooks]>=0.3.2",
    "ai-atlas-nexus[ollama]",
    "rich",
    "typer",
    "isort",
    "anthropic"
]

[project.optional-dependencies]
tests = ["pytest", "pytest-cov"]

[tool.pdm.version]
source = "file"
path = "src/mellea_skills_compiler/__init__.py"


[project.scripts]
mellea-skills = "mellea_skills_compiler.cli:app"

[tool.isort]
profile = "black"
line_length = 88
remove_redundant_aliases = true
group_by_package = true
combine_star = true
lines_after_imports = 2

[tool.pdm.build]
package-dir = "src"
